[GLUTEN-6110] Parallel run gluten ut and spark ut 1. Add ut notes 2. move page index test to gluten package
